+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
-## v1.0.0 - *Latest*
+## v1.0.1 - *Latest*
+- Changed initial settings check to see if "TFC_INSTANCE" is blank vs the built url
+- Added a check to see if there is a new release, and prompt for install.
+- Added possibility to add more buttons when making dialogs
+
+## v1.0.0
 - Initial Build of app
 
 ## Future Adds